+++ admin/modules/categoryProduct/delete.php
@@ -11,15 +11,32 @@ if(empty($EditCategory))
     $_SESSION['error'] = "Dữ liệu không tồn tại";
     redirectAdmin("categoryProduct");
 }
-$num = $db->delete("category_product", $id);
-if ($num > 0)
+/*
+ * Kiểm tra xem danh mục có sản phẩm
+ * */
+
+$is_product = $db->fetchOne("product", " category_id = $id ");
+//_debug($is_product);die;
+if ($is_product == NULL)
 {
-    $_SESSION['success'] = "Xóa thành công";
-    redirectAdmin("categoryProduct");
+    $num = $db->delete("category_product", $id);
+    if ($num > 0)
+    {
+        $_SESSION['success'] = "Xóa thành công";
+        redirectAdmin("categoryProduct");
+    }
+    else
+    {
+        $_SESSION['error'] = "Xóa thất bại";
+        redirectAdmin("categoryProduct");
+    }
 }
 else
 {
-    $_SESSION['error'] = "Xóa thất bại";
+    $_SESSION['error'] = "Danh mục có sản phẩm ! Bạn không thể xóa";
     redirectAdmin("categoryProduct");
 }
+
+
+
 ?>
